+++ ml_tools/ML_chain/_imprimir.py
@@ -2,6 +2,7 @@ from .._core import _imprimir_disponibles
 
 _GRUPOS = [
     "DragonChainOrchestrator",
+    "derive_next_step_schema",
     "augment_dataset_with_predictions",
     "augment_dataset_with_predictions_multi",
     "prepare_chaining_dataset",